code editted

+++ 0x0F-function_pointers/function_pointers.h
@@ -3,7 +3,7 @@
 
 void _putchar(void);
 void print_name(char *name, void (*f)(char *));
-void array_iterator(int *array, size_t size, void (*action)(int));
+void array_iterator(int *array, unsigned int size, void (*action)(int));
 int int_index(int *array, int size, int (*cmp)(int));
 #include <stdlib.h>
 #endif
